requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
joblib==1.5.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sniffio==1.3.1
starlette==0.47.3
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
